# Add intrinsics for floating point Math.min/Math.max on AArch64