--- a/compiler/src/org.graalvm.compiler.lir.aarch64/src/org/graalvm/compiler/lir/aarch64/AArch64ArithmeticOp.java
+++ b/compiler/src/org.graalvm.compiler.lir.aarch64/src/org/graalvm/compiler/lir/aarch64/AArch64ArithmeticOp.java
@@ -90,6 +90,8 @@ public enum AArch64ArithmeticOp {
     FRINTM,
     FRINTN,
     FRINTP,
+    FMAX,
+    FMIN,
     SQRT;
 
     /**
@@ -348,6 +350,12 @@ public enum AArch64ArithmeticOp {
                     break;
                 case FDIV:
                     masm.fdiv(size, dst, src1, src2);
+                    break;
+                case FMAX:
+                    masm.fmax(size, dst, src1, src2);
+                    break;
+                case FMIN:
+                    masm.fmin(size, dst, src1, src2);
                     break;
                 case MULVS:
                     masm.mulvs(size, dst, src1, src2);