Emit `fmin/fmax` instructions when `Math.min()/Math.max()` function is called with single or double precision floating point values on AArch64. The existing mechanism emits `cmp+csel` combination for the integer variants of `Math.min()/Math.max()`.

Relevant snippet of the microbenchmark is shown below (only the double variant for `Math.max()` is shown for brevity).

```
@Benchmark
public void jmhTimeMaxDouble() {
    double res = 0.0d;
    for (int i = 0; i < NUM_INVOKES; ++i) {
        double a = rand_double[i % NUM_RANDS];
        double b = rand_double[(i + 1) % NUM_RANDS];
        res += Math.max(a, b);
    }
    res_double = res;
}
```

Performance results on an A72 AArch64 machine

Before:
```
Benchmark         Mode  Cnt    Score   Error  Units
jmhTimeMaxDouble  avgt   15  288.019 ± 5.645  ns/op
jmhTimeMaxFloat   avgt   15  283.060 ± 4.367  ns/op
jmhTimeMinDouble  avgt   15  282.940 ± 3.823  ns/op
jmhTimeMinFloat   avgt   15  279.740 ± 2.278  ns/op
```

After:
```
Benchmark         Mode  Cnt   Score   Error  Units
jmhTimeMaxDouble  avgt   15  98.860 ± 0.010  ns/op
jmhTimeMaxFloat   avgt   15  98.847 ± 0.008  ns/op
jmhTimeMinDouble  avgt   15  98.867 ± 0.024  ns/op
jmhTimeMinFloat   avgt   15  98.855 ± 0.016  ns/op
```

Change-Id: Ic3b9de4aa096a8af699cfe5d0f0b8c83ca86b208